requirements.txt
amqp==5.1.1
asgiref==3.5.2
beautifulsoup4==4.11.1
celery==5.2.7
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
django-allauth==0.51.0
django-bootstrap4==1.1.1
django-celery-beat==2.3.0
django-elasticsearch-dsl>=6.0.0,<7.0.0
django-elasticsearch-dsl-drf==0.22.4
django-nine==0.2.5
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
elasticsearch>=6.0.0,<7.0.0
elasticsearch-dsl>=6.0.0,<7.0.0
filelock==3.7.1
idna==3.3
kombu==5.2.4
numpy==1.23.0
oauthlib==3.2.0
pandas==1.4.3
Pillow==9.1.1
pipenv==2022.6.7
platformdirs==2.5.2
prompt-toolkit==3.0.30
PyJWT==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz
requests==2.28.0
requests-oauthlib==1.3.1
simplejson
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
wincertstore==0.2
Django>=3.0,<4.0
psycopg2>=2.8
social-auth-app-django==5.0.0